+++ dotnet/src/StreamRecorder.Core/Models/StreamProtocol.cs
@@ -4,4 +4,5 @@ public enum StreamProtocol
 {
     Http = 0,
     Hls = 1,
+    Mmsh = 2,
 }